daniele_marco/daniele/mnt/user-data/outputs/contractiq/dspy_agents/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
dspy-ai==2.4.12
sqlalchemy==2.0.25
pydantic==2.5.3
httpx==0.26.0
redis==5.0.1
numpy<2.0.0
pyarrow<15.0.0